refactor: Pull bracket creation logic out of view controller
