+++ foam-madness/View/Game/GameScoreView.swift
@@ -39,7 +39,7 @@ struct GameScoreView: View {
             
             if isTournament {
                 Button("Back to Games", action: {
-                    NavigationUtil.popToTournamentGames(tournamentName: game.tournament?.name ?? "")
+                    NavigationUtil.popToViewByTitle(viewTitle: game.tournament?.name ?? "")
                 })
                 .buttonStyle(PrimaryButtonFullWidthStyle()).padding()
             } else {
